requirements.txt
streamlit
torch
torchvision
numpy
scikit-image
joblib
scikit-learn
Pillow
ultralytics
matplotlib
